+++ NEWS.md
@@ -1,6 +1,14 @@
 # amphidata 0.0.1
 
-* Initial CRAN submission.
+## Carrega de dades originals
+* Generació de dos objectes que contenen la totalitat de les dades del paquet.
+* Localització: `~/amphidata/data-raw/`.
+* Arxiu: `~/amphidata/R/sysdata.rda`.
+* Objectes: `l_df_ori` (matrius de dades Vázquez-Santiago) i `l_df_ori_88` (matrius de dades Golvin).
+
+## Funcions
+* `00_load_dimensions_vazquez()`: Funció per carregar (usa dades internes), seleccionar i filtrar les dades de Vázquez-Santiago.
+* `00_load_dimensions_golvin()`: Funció per carregar (usa dades internes), seleccionar i filtrar les dades de Golvin.
 
 # amphidata 0.0.0.9000